+++ lib/ui/res/app_assets.dart
@@ -1,0 +1,3 @@
+class AppAssets {
+  static const favorite = 'res/icons/heart_icon.svg';
+}
