+++ lib/ui/res/app_strings.dart
@@ -1,0 +1,3 @@
+class AppStrings {
+  static const appTitle = 'Список\nИнтересных мест';
+}
