+++ lib/ui/res/app_typography.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class AppTypography {
+  static const textText16Regular = TextStyle(
+    fontSize: 16.0,
+    height: 1.25,
+    fontWeight: FontWeight.w400,
+  );
+
+  static const textAppTitleStyle = TextStyle(
+    color: Colors.black,
+    fontWeight: FontWeight.bold,
+    fontSize: 30,
+  );
+}
