+++ lib/ui/screen/sight_details.dart
@@ -5,7 +5,6 @@ import 'package:flutter/src/widgets/framework.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/mocks.dart';
-import 'package:places/ui/res/constants.dart';
 
 // карточка с ПОДРОБНОЙ информацией о месте
